requirements.txt
Flask
Flask-Cors
flask-openapi3
Flask-SQLAlchemy
nose2
pydantic
SQLAlchemy
SQLAlchemy-Utils
typing_extensions
boto3
python-jose[cryptography]
requests
pudb